+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Updated dependencies. #INT-3324
+
+### Changed
+- Moved tinymce dependency to be a optional peer dependency. #INT-3324
+
 ## 8.0.1 - 2024-07-12
 
 ### Fixed