+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 9.0.0 - 2025-05-29
+
 ### Added
 - Added 'readonly' property. #TINY-11907